WEB(BE)/requirements.txt
#
# These requirements were autogenerated by pipenv
# To regenerate from the project's Pipfile, run:
#
#    pipenv lock --requirements
#

-i https://pypi.org/simple
aiofiles==0.6.0
click==7.1.2; python_version >= '2.7' and python_version not in '3.0, 3.1, 3.2, 3.3, 3.4'
fastapi==0.63.0
gunicorn==20.0.4
h11==0.12.0; python_version >= '3.6'
httptools==0.1.1
jinja2==2.11.3
markupsafe==1.1.1; python_version >= '2.7' and python_version not in '3.0, 3.1, 3.2, 3.3'
pydantic==1.7.3; python_version >= '3.6'
pymongo==3.12.0
python-dotenv==0.15.0
redis==3.5.3
starlette==0.13.6; python_version >= '3.6'
uvicorn==0.13.3
uvloop==0.14.0
motor==2.5.1
python-decouple==3.5

PyJWT
bcrypt
pydantic[email]
passlib
